hoomd_rs_vector -> hoomd_vector
