+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 
 members = [
-  "hoomd-rs-vector"
+  "hoomd-vector"
 ]
 resolver = "2"
