--- hoomd-rs-vector/README.md
@@ -1,5 +1,0 @@
-# hoomd-rs-vector
-
-Vector math library for [hoomd-rs].
-
-[hoomd-rs]: https://github.com/glotzerlab/hoomd-rs
