+++ hoomd-vector/README.md
@@ -1,0 +1,5 @@
+# hoomd-vector
+
+Vector math library for [hoomd].
+
+[hoomd-rs]: https://github.com/glotzerlab/hoomd-rs
